feat: revert route path param for image retrieval
This commit now uses `imageName` instead of `imageSha` to
retrieve the image. There is an extra API call with this change but it
allows the frontend to stick it to its default routing configuration.
The previous changes would have required the frontend to have its route
set to a path ending with the `imageSha` and not the `imageName`