[WIP]feat(menu): add route to display menu tree
This newly added route branches out all the relevant items that are
present in the first-level-nav (i.e pages, collections, etc.). What it basically does is read
`_data/navigation.yml`, maps it out, finds the relevant collection
folder and retrieves the collection pages in them, and finally, groups
all the collection pages into the respective subcollection (if the
filename of the collection page requires for it).

A reminder that this commit is a WIP, and the titles of the subcollection are not retrieved
because no file content is read, only folder listing is done. This only
branches out collections, NOT resource rooms.