fix: check thirdnav-pages were grouped properly
There was a bug where if the last entry in the `accumulator` was a
`thirdnav` object, collectionPages that were meant to be in a newly
created `thirdnav` object was instead added as a child to the pre-existing object.
This was due to the fact that there was not any checks to see if the
`thirdnav-page` was meant to be a child in that thirdnav object.