feat(nav): show thirdnav titles
Before this commit, the route would just group up the thirdnav pages
and label it as `subcollection` without retrieving the intended title of
the subcollection. Now it reads in the file required to set the
`thirdnav` title.

DO KEEP IN MIND that subcollection and thirdnav are used interchangeably
here, but this commit is meant to follow the terms that are standardized
in isomer. Subcollections are no longer used and will be referred to as
`thirdnav`s and it's child pages will be referred to as
`thirdnav-pages`. `collections` will still be referred to as
`collections`