Feat: add endpoint for drag and drop navigation menu (#27)
* feat: Create MenuType for reading _data/ files

This is the intended FileType for reading files under the `_data/`,
presumably those files that are for the footer and navigation menus

* [WIP]feat(nav): add route to display menu tree

This newly added route branches out all the relevant items that are
present in the first-level-nav (i.e pages, collections, etc.). What it basically does is read
`_data/navigation.yml`, maps it out, finds the relevant collection
folder and retrieves the collection pages in them, and finally, groups
all the collection pages into the respective subcollection (if the
filename of the collection page requires for it).

A reminder that this commit is a WIP, and the titles of the subcollection are not retrieved
because no file content is read, only folder listing is done. This only
branches out collections, NOT resource rooms.

* feat: choose branch ref for file CRUD operations

You can now choose to set the branch ref to point to another branch for
any File related operations. By default the branch ref is set to
`staging`

* feat(nav): show thirdnav titles

Before this commit, the route would just group up the thirdnav pages
and label it as `subcollection` without retrieving the intended title of
the subcollection. Now it reads in the file required to set the
`thirdnav` title.

DO KEEP IN MIND that subcollection and thirdnav are used interchangeably
here, but this commit is meant to follow the terms that are standardized
in isomer. Subcollections are no longer used and will be referred to as
`thirdnav`s and it's child pages will be referred to as
`thirdnav-pages`. `collections` will still be referred to as
`collections`

* Modify code to correctly display simple pages titles instead of file names

* Properly deslugify file names into their titles

* Separate function used to build the collections object into a menu-utils file

* Modify response to include unlinked collections and unlinked resource rooms

* Modify helper functions to take be compatible with "Unlinked Pages" as a special collection

* Modify data input into _.differenceBy

* Deslugify resource room name for menu

* fix: remove redundant file type

* format: remove unnecessary comments and code

* format: remove unncessary comments and code

* format: removed unnecessary code

* refactor: Create a Tree class to handle tree logic

* format: store variables as attributes in class instance

* format: remove unnecessary code