feat: send and receive only fields that the frontend requires
This commit limits the response being sent to the frontend to the
relevant fields (title, favicon, resources_name, colors) from the
config.yml file.

It also modifies the code to receive these same fields and make
updates to the config file as opposed to rewriting it completely.

This commit also switches out the base64 library for the js-base64
library for the Settings file as we had some problems with
encoding/decoding with this library in the past