feat: use footer.yml instead of social-media.yml (#60)
* feat: use footer.yml instead of social-media.yml

Currently, we are reading and social media data from the
_data/social-media.yml file and creating one of these files if they
do not exist. However, this is wrong - we don't actually use a
social-media.yml file in the Isomer V2 templates. Instead this
information is all found in the _data/footer.yml file. This commit fixes
that by retrieving information from and writing to the footer.yml file
instead.

This commit also modifies the Settings class so that when retrieving
the settings information, the api calls for both the config and footer
files are made concurrently using the Bluebird library.

* fix: cleanup misc. code

Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>