fix: allow collections to be empty in _config.yml
Currently, when faced with a repo which does not have any collections,
our Collection class' `list` method throws an error since it tries to
use Object.keys on a non-existent object attribute.

This commit fixes that by only performing the operation if the `collections`
attribute exists on the config object, otherwise an empty array is
returned. This allows us to accommodate repositories without any
collections.