feat: refactor route functions to use central error handler
Currently, every single route function is wrapped in a try-catch
block to catch any errors that occur in the function. This means that
we would need to write code for every route function to handle their
errors. Instead of doing so, we can simply remove all these try-catch
blocks and use a central error handler instead by wrapping each route
function with the route handler, which will catch any errors thrown
by the route function and forward it to the error handler.

This commit does the above by removing try-catch functions from every route
function and wrapping it with the route handler.