feat: adjust auth middleware function
Some endpoints, such as the authentication endpoint with GitHub OAuth,
do not require the user to have an access token. In such cases, we don't
want to try to extract a cookie and access token from a user request, since
might not have one.

This commit chooses the appropriate auth function (noVerify - or on verification -
for the login and logout endpoints, but verifyJwt for the rest of the
endpoints). Any route that is passed in but not found in our list of endpoints
will receive a 404 response with the message `Unauthorized for unknown route'.