feat: retrieve access token from req object
In the previous commit, we created a middleware layer which extracts
the user's github access token and attached it directly to the req
object. In this commit, we remove all the lines of code in the route
functions which do this, and instead retrieve the access token from the
req object.