feat: add endpoint for cookie validation
This commit introduces a GET /cookie-validation endpoint which allows
the user to check whether their JWT token stored inside their cookie,
with is used for authentication, is still valid. This is needed because
users might be on our CMS, go to another page, and come back to the CMS
when their token has expired. When that happens, we need a way to check
whether their token has expired before deciding what to display the user:
- if the token has expired, we can display the login page for them to login again
- otherwise, we can display the `Sites` layout