feat: change redirection URL to frontend
Previously, upon receiving the user's access token from GitHub (from
the OAuth callback - reference: https://docs.github.com/en/free-pro-team@latest/developers/apps/authorizing-oauth-apps)
we redirected the user to the '/sites' route on the frontend, since
we wanted that to be the first page they see when they login.

However, this has become a problem as we introduce more sophisticated routing
on the frontend. For example, we want to redirect all users from the '/'
path to the '/sites' path if they are logged in since we don't want to show
the login page if the user is already logged in. Therefore, the login state
becomes the key condition for redirection.

The main problem is that the login state was set to `true` (on the frontend) even before
we received the callback from GitHub. This caused a redirection to '/sites' by us,
not by the callback flow. Since this occurred before we received the access token
in the callback flow, naturally we would get a 401 error since we don't have the
access token necessary to retrieve GitHub resources, and this results in an auto-logout.

The workaround for this is that we introduce an intermediary page between
'/sites' and the login page, which is what the backend redirects the frontend to
after it completes the callback flow. On this intermediary page, we can
then set the login state to be true - the order is now correct since we
set login state to be true AFTER the callback flow is completed rather than
before. And then we can perform a redirect to '/sites'. Alternatively we could
perform the setting of state on the '/sites' page itself, but I figured it would
be cleaner to put that in a separate component.

Therefore, the main change in this commit is to change the redirection url to the
frontend from '/sites' to '/auth'.