Feat/centralized error handler (#63)
* feat: add route handler wrapper

    This commit adds a route handler wrapper to allow for centralized
    error handling. The route handler wraps around route functions and
    catches any errors thrown by the route function, and passes it on to
    the error handler

* feat: log errors in central error handler

* feat: refactor route functions to use central error handler

Currently, every single route function is wrapped in a try-catch
block to catch any errors that occur in the function. This means that
we would need to write code for every route function to handle their
errors. Instead of doing so, we can simply remove all these try-catch
blocks and use a central error handler instead by wrapping each route
function with the route handler, which will catch any errors thrown
by the route function and forward it to the error handler.

This commit does the above by removing try-catch functions from every route
function and wrapping it with the route handler.

* feat: create new AuthError class

This commit creates an AuthError class which is configured to have a
`status` attribute set to `401`. By using this error, we can simply throw
it and let our error handler can take care of sending the error response.

* refactor: relocate error handler to a middleware file

This commit relocates the error handler function to its own file in the
`middleware` folder

* fix: file name typo

* fix: properly serialize error

This commit uses the serialize-error package to properly serialize
errors into plain objects, since error objects don't always behave
like plan objects, which may make them impossible to stringify. We then
stringify the serialized error and console log it.

* feat: create a BaseIsomerError class

This commit creates a BaseIsomerError class which all other error
classes should be based on. This will allow us to distinguish our errors
from other unknown errors, so we can treat them differently in terms of
how we want to send the error response.

* feat: impose structure on http error response

This commit modifies the central error handler so that all error
responses have the following json structure:

{
  error: {
    name,
    code,
    message,
  }
}

and our own custom errors, which have an `isIsomerError` attribute,
will populate the name, code, and message attributes.

* feat: extract access token from jwt in middleware

This commit adds an auth router which extracts the access token from
the cookie, decrypts it, and adds it to the req object. By doing this
in the middleware layer, we don't have to do it in every single route
function, which is how it's being done right now.

* feat: adjust auth middleware function

Some endpoints, such as the authentication endpoint with GitHub OAuth,
do not require the user to have an access token. In such cases, we don't
want to try to extract a cookie and access token from a user request, since
might not have one.

This commit chooses the appropriate auth function (noVerify - or on verification -
for the login and logout endpoints, but verifyJwt for the rest of the
endpoints). Any route that is passed in but not found in our list of endpoints
will receive a 404 response with the message `Unauthorized for unknown route'.

* feat: retrieve access token from req object

In the previous commit, we created a middleware layer which extracts
the user's github access token and attached it directly to the req
object. In this commit, we remove all the lines of code in the route
functions which do this, and instead retrieve the access token from the
req object.

* feat: implement AuthError

* feat: implement NotFoundError

* fix: clean up code

This commit cleans up instances of code where I forgot to remove
the extraction of the JWT from the cookie.

* feat: rename cookie from `oauthtoken` to `isomercms`

* feat: cleanup code

* feat: add endpoint for cookie validation

This commit introduces a GET /cookie-validation endpoint which allows
the user to check whether their JWT token stored inside their cookie,
with is used for authentication, is still valid. This is needed because
users might be on our CMS, go to another page, and come back to the CMS
when their token has expired. When that happens, we need a way to check
whether their token has expired before deciding what to display the user:
- if the token has expired, we can display the login page for them to login again
- otherwise, we can display the `Sites` layout

* chore: rename cookie validation route to GET /validate-cookie

* feat: handle JWT errors

Previously, the `verifyJwt` auth function only handled jwt
`TokenExpiredError`s. But there are other types of authentication
errors possible, such as `JsonWebTokenError`. This commit accounts
for the `JsonWebTokenError` and also throws a generic error if the
error is not accounted for

* chore: typo

* chore: remove cookie validation endpoint

This commit removes all mention of the cookie validation endpoint, since
it is no longer needed by the frontend.

* feat: change redirection URL to frontend

Previously, upon receiving the user's access token from GitHub (from
the OAuth callback - reference: https://docs.github.com/en/free-pro-team@latest/developers/apps/authorizing-oauth-apps)
we redirected the user to the '/sites' route on the frontend, since
we wanted that to be the first page they see when they login.

However, this has become a problem as we introduce more sophisticated routing
on the frontend. For example, we want to redirect all users from the '/'
path to the '/sites' path if they are logged in since we don't want to show
the login page if the user is already logged in. Therefore, the login state
becomes the key condition for redirection.

The main problem is that the login state was set to `true` (on the frontend) even before
we received the callback from GitHub. This caused a redirection to '/sites' by us,
not by the callback flow. Since this occurred before we received the access token
in the callback flow, naturally we would get a 401 error since we don't have the
access token necessary to retrieve GitHub resources, and this results in an auto-logout.

The workaround for this is that we introduce an intermediary page between
'/sites' and the login page, which is what the backend redirects the frontend to
after it completes the callback flow. On this intermediary page, we can
then set the login state to be true - the order is now correct since we
set login state to be true AFTER the callback flow is completed rather than
before. And then we can perform a redirect to '/sites'. Alternatively we could
perform the setting of state on the '/sites' page itself, but I figured it would
be cleaner to put that in a separate component.

Therefore, the main change in this commit is to change the redirection url to the
frontend from '/sites' to '/auth'.

* feat: add hash string to redirect URL

This commit adds a hash string to the frontend redirection URL to add
context.

Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>