fea:t: remove unescaped characters to avoid 403 error on deploy
Currently, we get the following error when we attempt to deploy to
AWS EB with certain commit messages:

```
Deployment failed: Error: Status: 403. Code: SignatureDoesNotMatch,
Message: The request signature we calculated does not match the
signature you provided. Check your AWS Secret Access Key and signing
method. Consult the service documentation for details.
```

This is not because our access key is wrong, but because there are
certain characters in our commit message which are not dealt with
properly by the beanstalk-deploy package. Rather than wait for a fix
from the package author, we can first truncate the message, since the
title is in the first few words, to lower the chance of a special
character, and then we can replace these characters ourselves
in the deploy step when setting the version description.