feat: refine conditions for workflow jobs
This commit refines the conditions for the `gatekeep` and `test`
workflow jobs.

Firstly, we set `test` to run only if the github action type is
`pull_request` or if it's a `push` and the branch being pusehd to
is one of the pre-defined branches we've set as env vars in the
workflow file, since we don't want to waste build minutes running
on pointless commits in feature branches.

Secondly, we refine the `gatekeep` job to run only if the branch
is one of the branches specified in our env vars. This requires us
to create a new job, `exportEnv`, to retrieve the workflow-level env
vars as they are not available at the job-level: this is a missing
feature that has been referenced in the GitHub forums (link here:
https://github.community/t/how-to-set-and-access-a-workflow-variable/17335/15)