chore: remove cookie domain
On the staging version of the cms (staging-cms.isomer.gov.sg,
which hits our staging backend server staging-cms-api.isomer.gov.sg),
the frontend is sending the same cookie twice in the cookie header
which is causing the api call to fail because the backend can't read the cookie properly

I've set up an online dev environment mimicking the setup,
staging-dev-cms.isomer.gov.sg and staging-dev-cms-api.isomer.gov.sg, and
deployed the exact same code to them both, and set all configurations
to be the same and the cookie sends properly (appears only once), so i'm
not sure what the problem is the problem appears to be with the backend,
because i swapped the backends so that staging-cms.isomer.gov.sg points to
staging-dev-cms-api.isomer.gov.sg and vice-versa, and then the problem disappeared
on staging-cms and reappeared on staging-dev-cms

Yuanruo suggested that the problem is because our cookie domain is
set to our root domain, which means it will be sent by the browser on
subdomains, which might then result in key resolution issues. He
suggested removing the cookie entirely and letting it be set by
express.