chore: install dependencies needed for logger component
This commit installs the necessary dependencies for the logger
component. The reason we need a separate logger, which logs to
CloudWatch directly, is because the EC2 instance that our server
is running on might batches log statements before sending them to
CloudWatch. Logs might fail to get sent if the EC2 cloudwatch daemon
is corrupted or fails in some way. This external logger will send
logs instantly to CloudWatch so we have a fallback source of logs.

We chose to use the Winston logger as we have had success using it
for the homer project. this commit also installs the aws sdk as we
need to retrieve the EC2 instance id to log to the correct
cloudwatch log group.