feat: add singleton CloudWatchLogger class
This commit adds a CloudWatchLogger class which uses the singleton
pattern. Since a logger does not hold any state, it makes sense to
use the singleton pattern.

This logger attempts to connect to and log directly to cloudwatch
if in a production environment (NODE_ENV = prod or staging) and
logs directly to console if in dev mode. In other words, in a prod
environment we log both directly to cloudwatch and to console, so we
have two sets of logs as a fallback mechanism. In particular, our
cloudwatch only logs provide a clean application log which make it
easier to debug our own operations.