feat: add api-logging middleware
This commit adds an api logger middleware which logs all api requests
that have been submitted to the server for auditing purposes. Note that
this middleware is added AFTER the auth middleware instead of before -
that's because we're reliant on the auth middleware to add the user id
to the request object which will allow our api logs to be identifiable.