feat: add routes for creation and retrieval of third nav data
This commit adds two new routes:
- GET /sites/:siteName/collections/:collectionName/nav
- POST /sites/:siteName/collections/:collectionName/nav

These endpoints retrieve and create a yaml file in the _data folder
which describes a collection's pages and third nav sections. Each
collection has its own yaml file - the reason is so that we can have
more than one user working on the repo at the same time. The caveat
is that if they work on files in the same collection, there is still a
chance of a concurrency issue where two people working on the same repo
write conflicting versions of the file.

The functionality of the creation endpoint has been tested and completed
with this commit. The functionality of the GET endpoint will be
completed in subsequent commits.