fix: remove logging of request body
This commit modifies the API logger so that we log the request body.
The request body usually contains encrypted post content, and the relevant
file sha, which isn't important to us when we're sifting the logs.
Furthermore, we can identify the files being changes simply by looking
at the request path and date-time of the request - from these, we would
be able to identify the file being changed and also identify the file
version from the date-time of teh request.
that useful to us when we