fix: unprotect logout endpoint
Currently, the logout endpoint is protected by the `verifyJwt` auth
function. However, this prevents auto-logout functionality for when
a user's token expires, leaving them stuck in the UI with a useless
JWT which is incapable of making successful API calls.

This commit reverts the endpoint to be unprotected to fix this problem.