feat: update homepage title together with footer title
This commit resolves issue #269. If a user edits the title field,
they will update both the footer title and the browser title so that
the titles are consistent.

This commit achieves this by updating the title attribute of the homepage
file ONLY IF a change in the `title` field is detected.