Feat/add settings fields (#85)
* feat: add shareicon and analytics fields for Settings

* feat: retrieve and update agency logo in navigation.yml

This commit adds additional steps for the Settings route to retrieve
and update the navigation.yml file to update agency logo

* refactor: Settings class file retrieval and update process

This commit refactors the Settings class so that we can reuse a
utility function to retrieve the necessary settings files.

It also simplifies and standardizes the update process for all settings
files involved by putting the settings objects (config, navigation, footer)
in an array and looping over them.

* feat: update homepage title together with footer title

This commit resolves issue #269. If a user edits the title field,
they will update both the footer title and the browser title so that
the titles are consistent.

This commit achieves this by updating the title attribute of the homepage
file ONLY IF a change in the `title` field is detected.

* feat: send is_government field in response

This commit sends an additional is_government field when retrieving
settings configurations. The is_government field determines whether the
government masthead is displayed on the website.

* fix: error when updating settings object

* fix: retrieve settings configurations programmatically

Currently, we retrive the different settings configuration from
various files using an array. This is confusing since the settings files
are associated with random array indices. This commit switches to use
an object so that the operations to retrieve the files are associated
with a key instead.

* chore: replace arrays and magic indice numbers with objects

Previously, we used magic numbers to reference specific index
positions in arrays to retrieve data. This commit replaces such arrays
with objects so that data is accessible by key instead.

Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>