Fix: throw NotFoundError for files (#87)
This commit fixes the behaviour for throwing NotFoundErrors in File. Previously, it was not possible to reach the condition to throw the NotFoundError in read, as a file which could not be retrieved would have have a defined endpoint. This resulted in a different error being thrown. This commit fixes that and also adds in the check during deleting and updating files.