feat: only update Settings file if field has changed
The Settings class deals with three different types of configuration files:
_config.yml, _data/footer.yml, and _data/navigation.yml. Originally, we would
write to all three files, even if the only fields that were changed involved
only one of the files.

For example, even if I had only changed the `title` field (which affects only
the _config.yml file), I would end up making API calls to update the remaining
two settings files with their original content.

This commit allows us to save on wasted resources, and also increase the speed
of response by only updating a file if it needs to be updated.