feat: configure higher limit for nginx client_max_body_size
Even though we set a higher file size limit in our express app, we
found that uploads exceeding 1mb were being rejected by our server.
This was only occurring on the staging website, but not on local dev
env, which implies a server config issue.

After reading the nginx error logs, we found that nginx was rejecting
the requests because the body size of the request was too large. As a
result, we are introducing an ebextension which increases the max
body size (client_max_body_size) to 10mb (a slight increase over the 7mb
limit we set for express, since we want to be conservative and allow
for additional overhead).

Finally, the ebextension file includes a command to reload the nginx
service so that the new configs are used.