feat: add listAll method to File class
This commit adds a new `listAll` method to the File class. This method
returns data of both type `file` as well as type `dir` from GitHub. Previously,
the `list` method only returned data of type `file`. With the new
method, we have information on both files and directories, which we
need to render the nested sub-folders.

Since we already have a separate API for retrieving individual file content,
we need to limit the scope of this method - we should only retrieve content from
directories. If the provided endpoint is an individual file instead of a directory,
then we will throw a BadRequestError.