feat: add generalized endpoint for retrieving files within a folder or subfolder
This commit adds the GET /v1/sites/:siteName/folders endpoint which
retrieves the files and directories within the provided path directory.
This endpoint will fail if the provided path is not a valid directory (for
example, if you give it the path for an individual file, or if the provided
path does not exist in the repo) and will respond with the appropriate error.

The path is passed as a route query param `path` since the file/directory
path contains the `/` character, which prevents it from being used as a
route param.