refactor: reuse repo tree information when renaming collections
This commit attaches the `currentCommitSha` and `treeSha` data obtained
when using the `attachRollbackRouteHandlerWrapper` function and attaches
them to the request object. This allows us to reuse this info (for example,
when renaming collections) instead of having to retrieve them again.

This commit updates the `rename` method for the `Collections` class
to accept the `currentCommitSha` and `treeSha` as method arguments
instead of retrieving them again in a separate api call.