fix: collection deleting functionality
Previously, the `delete` method of the `Collection` class did not
work as expected - in particular, the method failed when we tried
to call `IsomerFile.delete` simultaneously on multiple files, with
the GitHub API responding with a 409 conflict error.

The reason why GitHub responds with a 409, is that when a file is
deleted from a directory, a new tree object (sans the deleted file) is
created, with a new tree SHA. The root/parent directory then points to
this new tree SHA. The conflict arises because when we try to delete files
concurrently, we end up referencing different tree objects, since the
deletion of an object changes the directory's tree SHA. Therefore, deletion
of files must be done serially, and not concurrently.

As a fix, instead of deleting the files one by one through the GitHub API,
we can use git trees to delete the entire collection instead. This
also simplifies things for us since we don't need to separately instantiate
a `CollectionConfig` class to delete the collection config file.