fix: use of `Collection.delete` in `deleteCollectionPage`
This commit makes two changes:

1. Update the usage of the `Collection.delete` method in the `deleteCollectionPage`
route function to include the additional params introduced (`currentCommitSha`
and `treeSha`).

2. Update the condition for checking whether to delete the collection.
Previously, we deleted the collection/folder directory if there were
no other pages left in the collection (or if `_.isEmpty(collectionPages)`).
However, with our migration to Jekyll 4, we now have a `collection.yml`
file remaining and so this `if` block will never be accessed. Therefore we
update the condition to check for whether the only file remaining is
the `collection.yml`.