Fix/delete collection conflict (#132)
* refactor: reuse repo tree information when renaming collections

This commit attaches the `currentCommitSha` and `treeSha` data obtained
when using the `attachRollbackRouteHandlerWrapper` function and attaches
them to the request object. This allows us to reuse this info (for example,
when renaming collections) instead of having to retrieve them again.

This commit updates the `rename` method for the `Collections` class
to accept the `currentCommitSha` and `treeSha` as method arguments
instead of retrieving them again in a separate api call.

* fix: collection deleting functionality

Previously, the `delete` method of the `Collection` class did not
work as expected - in particular, the method failed when we tried
to call `IsomerFile.delete` simultaneously on multiple files, with
the GitHub API responding with a 409 conflict error.

The reason why GitHub responds with a 409, is that when a file is
deleted from a directory, a new tree object (sans the deleted file) is
created, with a new tree SHA. The root/parent directory then points to
this new tree SHA. The conflict arises because when we try to delete files
concurrently, we end up referencing different tree objects, since the
deletion of an object changes the directory's tree SHA. Therefore, deletion
of files must be done serially, and not concurrently.

As a fix, instead of deleting the files one by one through the GitHub API,
we can use git trees to delete the entire collection instead. This
also simplifies things for us since we don't need to separately instantiate
a `CollectionConfig` class to delete the collection config file.

* fix: use of `Collection.delete` in `deleteCollectionPage`

This commit makes two changes:

1. Update the usage of the `Collection.delete` method in the `deleteCollectionPage`
route function to include the additional params introduced (`currentCommitSha`
and `treeSha`).

2. Update the condition for checking whether to delete the collection.
Previously, we deleted the collection/folder directory if there were
no other pages left in the collection (or if `_.isEmpty(collectionPages)`).
However, with our migration to Jekyll 4, we now have a `collection.yml`
file remaining and so this `if` block will never be accessed. Therefore we
update the condition to check for whether the only file remaining is
the `collection.yml`.

Co-authored-by: jiehao <jiehao@open.gov.sg>