style: use `prettier` plugin instead of `prettier-recommended`
This commit replaces the `plugin:prettier-recommended` with the
original `prettier` plugin. This is because our config does not
work with `plugin:prettier-recommended`.

The desired behavior when we run `npx eslint --fix .` is that all
linting bugs are fixed, but formatting issues are ignored. However,
we discovered that when we ran `npx eslint --fix .` with the
`prettier-recommended` plugin, formatting issues were beign corrected
as well. This is easily fixed by using the `prettier` plugin instead.