test: add alternate authentication scheme for e2e tests
This commit adds an alternate authentication scheme for e2e tests.
It checks for an E2E test secret, which is shared with the frontend,
and ensures that the repo being modified for these tests is only the
E2E test repo.

This is made possible by the implementation of Cypress on the frontend:
secrets associated with Cypress (such as COOKIE_NAME, or BASEURL) are
not bundled together with the deployment package, which means we are
free to use a shared secret string to authenticate the frontend with
the backend.

This commit also introduces new env vars for authentication of E2E
tests: E2E_TEST_REPO, and E2E_TEST_SECRET.