Fix: collectionPageService and subcollectionPageService tests
This commit fixes an issue in the subcollection page tests for create. We previously expected the same object to be returned in the convertDataToMarkdown function, which would always succeed regardless of the actual value of the third_nav_title parameter in the front matter, since the object itself was modified in the method.