refactor: update settings page
This commit refactors the update route for Settings. The refactor
simplifies the code structure by eliminating the use of an
intermediary "update" object which tracks which files need to be
updated. Using the new file services allows us to eliminate the
need for formatting the content before posting to GitHub.